backend/requirements.txt
fastapi
uvicorn
playwright
pandas
python-dotenv
openai
pydantic
beautifulsoup4
openpyxl
google-generativeai
crawl4ai
aiohttp
duckduckgo-search
